refactor: rename utils to internalUtils
