+++ src/internalUtils.ts
